[유저 관련 기능] 로그인 구현 및 수정, 삭제 권한 체크 (#35)
* feat(config) : ArgumentResolver 추가

* refactor(controller, service) : 수정, 삭제할 때 권한 체크

* feat(advice) : UserControllerExceptionAdvice 추가

* refactor(repository) : UserRepository.existsByEmail 삭제

+++ src/main/java/com/woowacourse/zzinbros/user/service/UserService.java
@@ -2,9 +2,13 @@ package com.woowacourse.zzinbros.user.service;
 
 import com.woowacourse.zzinbros.user.domain.User;
 import com.woowacourse.zzinbros.user.domain.UserRepository;
+import com.woowacourse.zzinbros.user.domain.UserSession;
 import com.woowacourse.zzinbros.user.dto.UserRequestDto;
+import com.woowacourse.zzinbros.user.exception.NotValidUserException;
 import com.woowacourse.zzinbros.user.exception.UserDuplicatedException;
+import com.woowacourse.zzinbros.user.exception.UserLoginException;
 import com.woowacourse.zzinbros.user.exception.UserNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,21 +21,30 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public User add(UserRequestDto userRequestDto) {
-        if (!userRepository.existsUserByEmail(userRequestDto.getEmail())) {
+    public User register(UserRequestDto userRequestDto) {
+        try {
             return userRepository.save(userRequestDto.toEntity());
+        } catch (DataIntegrityViolationException e) {
+            throw new UserDuplicatedException();
         }
-        throw new UserDuplicatedException();
     }
 
-    public User update(Long id, UserRequestDto userRequestDto) {
+    public User modify(Long id, UserRequestDto userRequestDto, UserSession userSession) {
         User user = findUser(id);
-        user.update(userRequestDto.toEntity());
-        return user;
+        if (userSession.matchEmail(user)) {
+            user.update(userRequestDto.toEntity());
+            return user;
+        }
+        throw new NotValidUserException("수정할 수 없습니다");
     }
 
-    public void delete(long id) {
-        userRepository.deleteById(id);
+    public void resign(Long id, UserSession userSession) {
+        User user = findUser(id);
+        if (userSession.matchEmail(user)) {
+            userRepository.deleteById(id);
+            return;
+        }
+        throw new NotValidUserException("삭제할 수 없습니다");
     }
 
     public User findUserById(long id) {
@@ -40,6 +53,21 @@ public class UserService {
 
     private User findUser(long id) {
         return userRepository.findById(id)
+                .orElseThrow(UserNotFoundException::new);
+    }
+
+    public UserSession login(UserRequestDto userRequestDto) {
+        User user = findUserByEmail(userRequestDto.getEmail());
+        User requestUser = userRequestDto.toEntity();
+
+        if (user.isAuthor(requestUser)) {
+            return new UserSession(userRequestDto.getName(), userRequestDto.getEmail());
+        }
+        throw new UserLoginException();
+    }
+
+    private User findUserByEmail(String email) {
+        return userRepository.findByEmail(email)
                 .orElseThrow(UserNotFoundException::new);
     }
 }
